Move JMX classes to a separate package
